+++ clients/browser_default.js
@@ -45,6 +45,7 @@ module.exports = {
   LexRuntime: require('./lexruntime'),
   MachineLearning: require('./machinelearning'),
   MarketplaceCommerceAnalytics: require('./marketplacecommerceanalytics'),
+  MechanicalTurkRequester: require('./mechanicalturkrequester'),
   MobileAnalytics: require('./mobileanalytics'),
   OpsWorks: require('./opsworks'),
   Polly: require('./polly'),