+++ clients/dynamodbstreams.js
@@ -8,6 +8,7 @@ AWS.DynamoDBStreams = Service.defineService('dynamodbstreams', ['2012-08-10']);
 Object.defineProperty(apiLoader.services['dynamodbstreams'], '2012-08-10', {
   get: function get() {
     var model = require('../apis/streams.dynamodb-2012-08-10.min.json');
+    model.paginators = require('../apis/streams.dynamodb-2012-08-10.paginators.json').pagination;
     return model;
   },
   enumerable: true,