+++ clients/mechanicalturkrequester.js
@@ -1,0 +1,18 @@
+require('../lib/node_loader');
+var AWS = require('../lib/core');
+var Service = require('../lib/service');
+var apiLoader = require('../lib/api_loader');
+
+apiLoader.services['mechanicalturkrequester'] = {};
+AWS.MechanicalTurkRequester = Service.defineService('mechanicalturkrequester', ['2017-01-17']);
+Object.defineProperty(apiLoader.services['mechanicalturkrequester'], '2017-01-17', {
+  get: function get() {
+    var model = require('../apis/mturk-requester-2017-01-17.min.json');
+    model.paginators = require('../apis/mturk-requester-2017-01-17.paginators.json').pagination;
+    return model;
+  },
+  enumerable: true,
+  configurable: true
+});
+
+module.exports = AWS.MechanicalTurkRequester;