+++ clients/waf.js
@@ -8,6 +8,7 @@ AWS.WAF = Service.defineService('waf', ['2015-08-24']);
 Object.defineProperty(apiLoader.services['waf'], '2015-08-24', {
   get: function get() {
     var model = require('../apis/waf-2015-08-24.min.json');
+    model.paginators = require('../apis/waf-2015-08-24.paginators.json').pagination;
     return model;
   },
   enumerable: true,